requirements.txt
python-dotenv
Numpy
pandas
scikit-learn
matplotlib
seaborn
pymongo
certifi
pymongo[srv]
